Refactor CtrfReportEngine into focused partial files (#9054)
Splits the 778-line CtrfReportEngine.cs into focused partial-class files, each under 300 lines, with no public/internal API change:

- CtrfReportEngine.cs - orchestration (constructor, GenerateReportAsync, GenerateReportCoreAsync)
- CtrfReportEngine.FileWriter.cs - collision-safe file persistence
- CtrfReportEngine.FileNaming.cs - output-path construction and sanitization
- CtrfReportEngine.JsonSerializer.cs - top-level CTRF JSON document
- CtrfReportEngine.JsonTestWriter.cs - per-test/attempt JSON writing
- CtrfReportEngine.TestCollapsing.cs - retry-attempt collapsing and output-line helpers
